backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.6.0
requests>=2.31.0
websocket-client>=1.7.0
python-dotenv>=1.0.1
pytest>=7.4.0
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
